requirements.txt
pyyaml
tqdm
numpy
easydict==1.9.0
joblib
matplotlib
pandas
albumentations==0.5.2
hydra-core==1.1.0
pytorch-lightning==1.2.9
tabulate
webdataset
packaging
scikit-learn==0.24.2
wldhx.yadisk-direct
opencv-python==4.5.5.64
